Changes to the way we're handling keying and how we're handling password upgrades...
We generate the key once and store it in it's own custom option now.
(We also remove it on plugin delete).

We upgrade the password during creation of the settings object if
the password doesn't contain one of our crypto type tags to identify
it as being one of the three ways we allow crypto (NACL, OpenSSL + GCM,
OpenSSL + HMAC).